Fixed bug in DPHiveJobStep.scala
